Разделил файлы интерфейса и файлы реализации.
