+++ SongStruct.h
@@ -1,0 +1,12 @@
+#ifndef STRUCT_H
+#define STRUCT_H
+
+const int AMOUNT_OF_SONGS = 20;
+
+struct Song {
+	std::string track;
+	std::string band;
+	int popularity;
+};
+
+#endif STRUCT_H
